refact: context in separate file
